requirements.txt
appdirs==1.4.4
asgiref==3.3.0
black==20.8b1
certifi==2020.11.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.8
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
dataclasses==0.8
Deprecated==1.2.13
Django==3.1.3
django-debug-toolbar==3.2.2
django-oauth-toolkit==1.5.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jwcrypto==1.0
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==20.4
pathspec==0.8.1
permission==0.4.1
Pillow==8.3.2
pycparser==2.21
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2020.4
regex==2020.10.28
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.13.3
zipp==3.6.0
